refactor: move integration tests
