+++ internal/cli/navigation/parent_test.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/stretchr/testify/require"
 
-	"stackit.dev/stackit/internal/cli/testhelper"
+	"stackit.dev/stackit/internal/testhelper"
 	"stackit.dev/stackit/testhelpers"
 )
