+++ internal/integration/parallel_metadata_test.go
