+++ internal/integration/staleness_repro_test.go
